refactor: Home routing
