+++ src/pages/Home/CommonWalkingMoneyRoad.tsx
@@ -1,0 +1,5 @@
+function CommonWalkingMoneyRoad() {
+  return <>걷고있는 돈길 상세</>;
+}
+
+export default CommonWalkingMoneyRoad;
