--- src/pages/Home/ProceedingChallenge.tsx
@@ -1,5 +1,0 @@
-function ProceedingChallenge() {
-  return <>부모-진행중인 챌린지</>;
-}
-
-export default ProceedingChallenge;
