+++ src/pages/Home/kid/KidHome.tsx
